Bug fix:
(1)An atom connected with two or more O atoms of the delocated bonds should not be a chiral center. But RDKit gives two O atoms different CIP ranks, which makes that atom to be a chiral center.

(2) Ligand names are kept to be the same upper/low cases as input string.